requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
kaggle>=1.5.13
pillow>=9.5.0
matplotlib>=3.7.1
tqdm>=4.65.0
